machine: Support an alternative kernel path
ArchLinux stores the kernel in `/boot/vmlinuz-linux` (assuming you
installed the package `linux`).

Note that if one installs the LTS kernel, the name will be
`vmlinuz-linux-lts`, and if one installs a custom kernel, the name will
be something like `vmlinuz-<whatever>`.

A better patch would be to look for /boot/vmlinux-*, and maybe error out
if more than one kernel is found.

Hence this commit is not the universal solution yet.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>