machine: Refine the logic to find the kernel release
We still pickup the last kernel found in /usr/lib/modules, however we
add an additional constraint: we expect the kernel release to start with
a digit. In effect, it allows to filter out other directories that might
be living there, for example the extramodules directory on ArchLinux.

    $ ls -1 /usr/lib/modules
    5.2.4-arch1-1-ARCH
    extramodules-ARCH

(this trick was taken from mkosi source code)

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>