machine: Support xz-compressed kernel modules
This is needed to run on ArchLinux, which ships xz-compressed kernel
modules.

Note that we also remove error checking from CopyFile(), first because
this function panics if the input is not found anyway, and second to be
consistent with the rest of the code, which never ever checks the return
value of this function.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>